+++ replicante/src/tasks/cluster_refresh/metrics.rs
@@ -1,28 +1,23 @@
 use lazy_static::lazy_static;
-use prometheus::CounterVec;
+use prometheus::Counter;
+use prometheus::Histogram;
 use prometheus::HistogramOpts;
-use prometheus::HistogramVec;
-use prometheus::Opts;
 use prometheus::Registry;
 use slog::debug;
 use slog::Logger;
 
 lazy_static! {
-    pub static ref REFRESH_DURATION: HistogramVec = HistogramVec::new(
+    pub static ref REFRESH_DURATION: Histogram = Histogram::with_opts(
         HistogramOpts::new(
             "replicore_cluster_refresh_duration",
             "Duration (in seconds) of a cluster state refresh task"
         )
         .buckets(vec![0.25, 0.5, 1.0, 2.5, 5.0, 10.0, 20.0, 40.0]),
-        &["cluster"]
     )
     .expect("Failed to create REFRESH_DURATION");
-    pub static ref REFRESH_LOCKED: CounterVec = CounterVec::new(
-        Opts::new(
-            "replicore_cluster_refresh_locked",
-            "Number of times a cluster refresh task was abandoned because the cluster is locked"
-        ),
-        &["cluster"]
+    pub static ref REFRESH_LOCKED: Counter = Counter::new(
+        "replicore_cluster_refresh_locked",
+        "Number of times a cluster refresh task was abandoned because the cluster is locked"
     )
     .expect("Failed to create REFRESH_LOCKED");
 }